feat: ikon sistemi, PV/CV ondalık desteği ve arayüz düzeltmeleri
İKON SİSTEMİ (Lucide, sıfır kütüphane bağımlılığı)
- Tüm ikonlar Lucide'a taşındı; lucide-react ve @svgr/webpack kaldırıldı
- İkonlar derin ESM import ile yerel modüle bağlandı (tree-shaking doğrulandı:
  kullanılmayan ikonlar pakete girmiyor)
- tools/ altında bağımlılıksız üretici + eşleme tablosu (yeni ikon kolayca eklenir)
- 58 SVG dosyası, 47 inline SVG ve ~157 ölü varlık temizlendi
- Marka işaretleri (Google/Facebook/LinkedIn/Instagram/X) yerel bileşen olarak

PV/CV ONDALIK DESTEĞİ (uçtan uca)
- Migration 0020: 21 PV/CV sütunu bigint -> numeric(14,2)
- Backend: 44 model alanı ve API DTO'ları float64, hesaplarda 2 hane yuvarlama
- Panel formlarında step=0.01, gösterimde formatPV (100 / 25,5 / 12,75)

KRİTİK HATA DÜZELTMESİ
- Panelden sipariş "ödendi" işaretlendiğinde ProcessOrderEffects çağrılmıyordu;
  PV/CV birikimi, referans primi, perakende kazancı ve binary ağaç güncellemesi
  hiç uygulanmıyordu. Artık durum "paid" olduğunda etkiler idempotent şekilde
  uygulanıyor (canlıda tüm kullanıcıların PV/CV'si bu yüzden 0 görünüyordu).

ARAYÜZ DÜZELTMELERİ
- account/tree: responsive sığdırma yeniden yazıldı (kart kırpılması giderildi,
  okunabilir ölçek, tuval yüksekliği ekrana göre, "Sığdır" gerçekten sığıyor)
- Sepet: mobilde iki satırlı düzen, yatay taşma sıfır
- PV mor / CV yeşil renk kuralı tüm sayfalarda (site + panel)
- Panelde ürün satırına tıklayınca "Düzenle" modal olarak açılıyor
- Üst menü aksiyonları simetrik (40px kutu, 22px ikon, aynı dikey merkez)

KİMLİK / TİPOGRAFİ
- Site geneli font Open Sans (flptr.com ile aynı)
- Başlıklar hariç tüm metinler normal ağırlıkta
- Logo Quicksand 800 + koyu ton (#101828) olarak iki tarafta eşitlendi

YEREL GELİŞTİRME
- local-start.sh / local-stop.sh (api + panel + eshop)
- Panel için dev-only /api proxy (oturum çerezi 401 sorunu)
- 3000/bestmanager -> panel yönlendirmesi, kök ve eski ad yönlendirmeleri
- icon.md: envanter, eşleme tabloları ve geçiş raporu

+++ mlm-backend/internal/handlers/admin_handler_extended.go
@@ -371,8 +371,8 @@ func (h *AdminHandler) AdjustPVAndCV(c *gin.Context) {
 		return
 	}
 	var req struct {
-		DeltaPV int64  `json:"delta_pv"`
-		DeltaCV int64  `json:"delta_cv"`
+		DeltaPV float64  `json:"delta_pv"`
+		DeltaCV float64  `json:"delta_cv"`
 		Reason  string `json:"reason" binding:"required"`
 	}
 	if err := c.ShouldBindJSON(&req); err != nil {
