+++ mlm-backend/internal/models/order_item.go
@@ -7,6 +7,6 @@ type OrderItem struct {
 	ProductID *int64  `json:"product_id"`
 	Quantity  int     `json:"quantity"`
 	Price     float64 `json:"price"`
-	PV        int64   `json:"pv"`
-	CV        int64   `json:"cv"`
+	PV        float64   `json:"pv"`
+	CV        float64   `json:"cv"`
 }
