+++ mlm-backend/internal/models/product.go
@@ -7,8 +7,8 @@ type Product struct {
 	ID           int64     `json:"id"`
 	Name         string    `json:"name"`
 	Price        float64   `json:"price"`
-	PV           int64     `json:"pv"`
-	CV           int64     `json:"cv"`
+	PV           float64     `json:"pv"`
+	CV           float64     `json:"cv"`
 	Stock        int       `json:"stock"`
 	Description  *string   `json:"description"`
 	ImagePath    *string   `json:"image_path"`
